Better errors in the CLI and simplified error handling
Basically, urfave/cli just doesn't allow us to hook our own exit logic
into the places that I need to be able to.  It is detecting things
correctly, but I can't intercept it.  Because of that, I'll remove the
ErrBadUsage condition and just have a special ErrCorrupt, which I have
unit tests to show is set correctly when the error is a corruption issue